Revert "HPIS-91 photo from camera"
