+++ .gitignore
@@ -1,4 +1,3 @@
 .gradle
 .idea
 build
-server/src/main/resources/secret.json
