+++ frontend/.gitignore
@@ -14,4 +14,4 @@ web-build/
 .DS_Store
 
 package-lock.json
-constants/
+constants/serverAddress.js
