+++ frontend/actions/login.js
@@ -6,20 +6,6 @@ export const SIGN_IN = 'SIGN_IN';
 export const REGISTER = 'REGISTER';
 export const CHECK_LOGIN = 'CHECK_LOGIN';
 
-async function fetchWithTimeout(resource, options = {}) {
-    const { timeout = 2000 } = options;
-    
-    const controller = new AbortController();
-
-    const timer = setTimeout(() => controller.abort(), timeout);
-    const response = await fetch(resource, {
-      ...options,
-      signal: controller.signal  
-    });
-    clearTimeout(timer);
-    return response;
-  }
-
 export const signIn = (login, password) => {
     return async dispatch => {
         console.log('signing in...')
@@ -27,31 +13,19 @@ export const signIn = (login, password) => {
         const name = login;
         password = SHA256(password).toString();
         
-        var response;
-
-        try {
-            response = await fetchWithTimeout(
-                `http://${serverAddress.address}:8080/login`,
-                {
-                    method: 'POST',
-                    headers: {
-                        'Content-type': 'application/json',
-                    },
-                    body: JSON.stringify({
-                        name,
-                        password
-                    })
-                }
-            );
-        } catch (error) {
-            if (error.name === 'AbortError') {
-                console.log("AbortError occurred");
-                return dispatch({ type: SIGN_IN, token: null });
+        const response = await fetch(
+            `http://${serverAddress.address}:8080/login`,
+            {
+                method: 'POST',
+                headers: {
+                    'Content-type': 'application/json',
+                },
+                body: JSON.stringify({
+                    name,
+                    password
+                })
             }
-            console.log("error occurred: " + error.toString());
-            return dispatch({ type: SIGN_IN, token: null });
-        }
-
+        );
         console.log(`response status: ${response.status}`);
         let token = null;
         if(response.status == 200) {
@@ -82,7 +56,6 @@ export const logInWithGoogle = googleToken => {
         });
         console.log(`response status: ${first_response.status}`);
         const respData = await first_response.json();
-        console.log(respData);
 
         const  { email } = respData;
         const { sub } = respData;
@@ -93,31 +66,19 @@ export const logInWithGoogle = googleToken => {
         const name = email;
         const password = SHA256(sub).toString();
         
-        var second_response;
-
-        try {
-        second_response = await fetchWithTimeout(
-                `http://${serverAddress.address}:8080/loginWithGoogle`,
-                {
-                    method: 'POST',
-                    headers: {
-                        'Content-type': 'application/json',
-                    },
-                    body: JSON.stringify({
-                        name,
-                        password
-                    })
-                }
-            );
-        } catch (error) {
-            if (error.name === 'AbortError') {
-                console.log("AbortError occurred");
-                return dispatch({ type: SIGN_IN, token: null });
+        const second_response = await fetch(
+            `http://${serverAddress.address}:8080/loginWithGoogle`,
+            {
+                method: 'POST',
+                headers: {
+                    'Content-type': 'application/json',
+                },
+                body: JSON.stringify({
+                    name,
+                    password
+                })
             }
-            console.log("error occurred: " + error.toString());
-            return dispatch({ type: SIGN_IN, token: null });
-        }
-
+        );
         console.log(`response status: ${second_response.status}`);
         let token = null;
         if(first_response.status == 200) {
@@ -141,30 +102,19 @@ export const register = (login, password) => {
         const name = login;
         password = SHA256(password).toString();
 
-        var response;
-        try {
-        response = await fetch(
-                `http://${serverAddress.address}:8080/register`,
-                {
-                    method: 'POST',
-                    headers: {
-                        'content-type': 'application/json',
-                    },
-                    body: JSON.stringify({
-                        name,
-                        password
-                    })
-                }
-            );
-        } catch (error) {
-            if (error.name === 'AbortError') {
-                console.log("AbortError occurred");
-                return dispatch({ type: REGISTER, outcome: null });
+        const response = await fetch(
+            `http://${serverAddress.address}:8080/register`,
+            {
+                method: 'POST',
+                headers: {
+                    'content-type': 'application/json',
+                },
+                body: JSON.stringify({
+                    name,
+                    password
+                })
             }
-            console.log("error occurred: " + error.toString());
-            return dispatch({ type: REGISTER, outcome: null });
-        }
-
+        );
         console.log(`response status: ${response.status}`);
         if(response.status == 201) {
             console.log(`Registration successful!`);
