+++ frontend/package.json
@@ -34,10 +34,7 @@
     "react-navigation": "^4.4.4",
     "react-redux": "^7.2.6",
     "redux": "^4.1.2",
-    "redux-thunk": "^2.4.1",
-    "expo-image-picker": "~11.0.3",
-    "expo-permissions": "~13.0.3",
-    "expo-camera": "~12.0.3"
+    "redux-thunk": "^2.4.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
