+++ frontend/screens/login/SignInScreen.js
@@ -1,6 +1,8 @@
 import React, {useDebugValue, useEffect, useState} from 'react';
 import { View, StyleSheet, Text, TextInput, Image, useWindowDimensions, ScrollView, ImageBackground, Pressable, TouchableOpacity } from 'react-native';
 import GoogleSignInButton from '../../components/GoogleSignInButton';
+//import GoogleButton from 'react-google-button'
+//import SignInWithGoogle from '../../components/SignInWithGoogle';
 import { useNavigation } from '@react-navigation/core';
 import Card from '../../components/Card';
 import OverscreenModal from '../../components/OverscreenModal';
