+++ server/settings.gradle.kts
@@ -1,2 +1,1 @@
 rootProject.name = "hello"
-include("java")
