+++ server/src/main/kotlin/pishello/hello/api/AuthEndpoint.kt
@@ -3,17 +3,17 @@ package pishello.hello.api
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import pishello.hello.persistence.database.entities.Token
-import pishello.hello.persistence.database.ports.TokenPort
-import pishello.hello.persistence.database.ports.UserPort
+import pishello.hello.database.Token
+import pishello.hello.database.TokenConnection
+import pishello.hello.database.UserConnection
 
 data class LoginRequest(val name: String, val password: String)
 
 @RestController
-class AuthEndpoint(val userPort: UserPort, val tokenPort: TokenPort) {
+class AuthEndpoint(val userConnection: UserConnection, val tokenConnection: TokenConnection) {
     @GetMapping("/check")
     fun check(@RequestParam name: String): ResponseEntity<Unit> {
-        return if (userPort.checkIfUserExists(name)) {
+        return if (userConnection.checkIfUserExists(name)) {
             ResponseEntity(HttpStatus.CONFLICT)
         } else
             ResponseEntity(HttpStatus.OK)
@@ -21,8 +21,8 @@ class AuthEndpoint(val userPort: UserPort, val tokenPort: TokenPort) {
 
     @PostMapping("/register", consumes = ["application/json"])
     fun register(@RequestBody request: LoginRequest): ResponseEntity<Unit> {
-        return if (!userPort.checkIfUserExists(request.name)) {
-            userPort.createNewUser(request.name, request.password)
+        return if (!userConnection.checkIfUserExists(request.name)) {
+            userConnection.createNewUser(request.name, request.password)
             ResponseEntity(HttpStatus.CREATED)
         } else
             ResponseEntity(HttpStatus.CONFLICT)
@@ -30,9 +30,9 @@ class AuthEndpoint(val userPort: UserPort, val tokenPort: TokenPort) {
 
     @PostMapping("/login", consumes = ["application/json"])
     fun login(@RequestBody request: LoginRequest): ResponseEntity<Token> {
-        val user = userPort.checkLogin(request.name, request.password)
+        val user = userConnection.checkLogin(request.name, request.password)
         return if (user != null) {
-            val token = tokenPort.createNewToken(user)
+            val token = tokenConnection.createNewToken(user)
             ResponseEntity(token, HttpStatus.OK)
         } else
             ResponseEntity(HttpStatus.FORBIDDEN)
@@ -40,13 +40,13 @@ class AuthEndpoint(val userPort: UserPort, val tokenPort: TokenPort) {
 
     @PostMapping("/loginWithGoogle", consumes = ["application/json"])
     fun google(@RequestBody request: LoginRequest): ResponseEntity<Token> {
-        return if (!userPort.checkIfUserExists(request.name)) {
-            val user = userPort.createNewUser(request.name, request.password)
-            ResponseEntity(tokenPort.createNewToken(user), HttpStatus.OK)
+        return if (!userConnection.checkIfUserExists(request.name)) {
+            val user = userConnection.createNewUser(request.name, request.password)
+            ResponseEntity(tokenConnection.createNewToken(user), HttpStatus.OK)
         } else {
-            val user = userPort.checkLogin(request.name, request.password)
+            val user = userConnection.checkLogin(request.name, request.password)
             if (user != null) {
-                ResponseEntity(tokenPort.createNewToken(user), HttpStatus.OK)
+                ResponseEntity(tokenConnection.createNewToken(user), HttpStatus.OK)
             } else
                 ResponseEntity(HttpStatus.FORBIDDEN)
         }
