+++ server/src/main/kotlin/pishello/hello/database/DataSource.kt
@@ -1,4 +1,4 @@
-package pishello.hello.persistence.database
+package pishello.hello.database
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
