+++ server/src/main/kotlin/pishello/hello/database/TokenRepository.kt
@@ -1,0 +1,8 @@
+package pishello.hello.database
+
+import org.springframework.data.repository.CrudRepository
+
+
+interface TokenRepository: CrudRepository<Token?, String?> {
+    fun findByName(name: User): Token?
+}
