--- server/src/main/kotlin/pishello/hello/persistence/database/repositories/TokenRepository.kt
@@ -1,10 +1,0 @@
-package pishello.hello.persistence.database.repositories
-
-import org.springframework.data.repository.CrudRepository
-import pishello.hello.persistence.database.entities.Token
-import pishello.hello.persistence.database.entities.User
-
-
-interface TokenRepository: CrudRepository<Token?, String?> {
-    fun findByName(name: User): Token?
-}
