+++ server/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.jpa.database-platform = pishello.hello.persistence.database.SQLDialect
+spring.jpa.database-platform = pishello.hello.database.SQLDialect
 spring.jpa.hibernate.ddl-auto = none
 
 spring.datasource.url = jdbc:sqlite:src/main/resources/main.db
